Get the "runtime" configuration to be found, and bump to v0.2.2-SNAPSHOT
v0.2.1 had a problem that loading it with the plugin DSL (like below) caused:
`org.gradle.api.artifacts.UnknownConfigurationException: Configuration with name 'runtime' not found.`

```
plugins {
    id "java"
    id "maven"
    id "org.embulk.embulk-plugins" version "0.2.1"
}
```

It was because the "runtime" configuration is defined in the "java" plugin, and
the plugin DSL does not apply the plugins immediately. This commit fixes the
problem by explicitly applying the "java" plugin.