Change "embulkPluginFlatRuntime" to "embulkPluginRuntime", and to be non-configurable
The configuration "flatRuntimeConfiguration" in "embulkPlugin" was not working.

Values of EmbulkPluginExtension can be retrieved only in afterEvaluate. On the other hand,
conf2ScopeMappings must be configured before evaluation (not in afterEvaluate). As a result,
it is impossible to configure the Configuration name in "embulkPlugin".

This commit gives up to configure it, and gives a fixed name "embulkPluginRuntime".
"flat" is removed there because the alternative configuration may have another purpose
in the future.

In addition, it turned out that the POM dependency modification works only for
Upload tasks, not in "maven-publish", because "maven-publish" does not use
"conf2ScopeMappings".