Make "mainJar" configurable so that another task than the normal "jar" task
Some Embulk plugins need to embed its dependency libraries in itself,
possibly with renaming/relocating their packages, for example with
the "shadow" Gradle plugin.
https://plugins.gradle.org/plugin/com.github.johnrengelman.shadow

The "shadow" Gradle plugin creates another "shadowJar" task in addition
to the default "jar" task. But, this "org.embulk.embulk-plugins" Gradle
plugin accepted only the default "jar" task. It couldn't be used with
the "shadow" Gradle plugin.

This change would mitigate the situation. It allows Embulk plugin
developers to build an Embulk plugin with the "shadow" Gradle plugin.
It requires "build.gradle" to be a little bit non-intuitive, though.