Unsupport "uploadArchives" with the "maven" plugin to delay creating the "embulkPluginRuntime" configuration after evaluation
The "embulkPluginRuntime" configuration needs to be configured after evaluation so that
it can have an Attribute to work with Gradle 6. See #71 for the details.

At the same time, Gradle started to deprecate "uploadArchives" with the "maven" plugin
from Gradle 6 to be possibly removed in Gradle 7. It's good time to stop supporting
"uploadArchives" even in this Gradle plugin.
https://github.com/gradle/gradle/issues/3003#issuecomment-495025844